Telemetry refactor (#114)
* IP and network information gathering finished. TODO: unittests and inplace tests

* Added unittests for all IP gatherings but swarm data. Also refactored the way of retrieving local interfaces IP from address to route.

* Locally fixes tests

* set IP tests updated

* Fixed code smells and updated test. Hardcoded IP are only on tests

* Added functionality and tests for Swarm IP gathering. Still using old function from Runtime.

* Added fix for sonar errors on hardcoded IP addresses

* Change random IP address generator to be cryptographically protected.

* Change auxiliary docker image to latest to prevent version conflicts.

* updates and refactors

* Network interface monitor renamed and re-located. Added base data structure to monitor module and sample edge_status.py module and class.

* Code clean and test refactored.

* Update test_network_interface_monitor.py

Deleted unnecesary comment

* minior: cosmetic changes

* Added monitor component auto importing, initialization and update. Added class for basic NuvlaEdge configuration data and its corresponding monitor.

* Added all components as monitors. They automatically run if used the flag @monitor before the class definition. UnitTest WIP

* Jetson nano tested. Power readings are intermittent and faulty

* Unit tests created for every component in monitor class.

* UnitTests for Vulnerabilities and Geolocation finished

* Fixing Network unittests

* Added UnitTests for traffic data readings in Network monitor

* Added Unittests for NuvlaEdge info and container stats monitors

* Added unittests for Resource Monitor class

* All monitors unittests completed. Pylint, and final in-board tests remaining

* Vulnerabilities fixed.

* Telemetry unittests fixed

* Changed attribute assignation in  for tests

* Changed attribute assignation in  for tests

* Removed autospec from patch in temperature monitor tests

* Temperature tests restructured

* Update main.yml

Tests for github action on docker registry tag creation

* Update main.yml

* Update main.yml

Restored previous status of GH actions workflow

* Update container-release.sh

Updated ccontainer release with an exception for pull request trigger actions

* Update container-release.sh

Fixed bug on previous commit

* Code smell review

* Fixed a bug where Nuvla requests to remove an empty peripheral and created an unhandled exception

* Added Telemetry.py class fix

* Added CHANGELOG.md from previous telemetry refactor

* Added thread recovery watchdog

* Added logging formatting for Agent component.
Added the possibility to disable monitors.

* Main app script refactored and working. Renamed files to match Python standards

* Removed app.py script in favor of agent_main.py. Dockerfile adapted to this change

* Added initial logger configuration file.

* Security hotspot endpoint API fixed

* Security hotspot endpoint API fixed V2

* Added executable line to main script

* RollBack to manager peripheral original

* Removed Agent Class tests

* Updated container memory report. From RSS+cache to RSS memory

* Fixed docker tests after refactor

* Fixed k8 tests after refactor

* Fixed NuvlaBoxCommon tests

* Clean some code smells

* Fixed tests after refactoring most of the exception handling

* Added container telemetry cleanup after sending heartbeat

* Agent class tests finished.

* Fixed imports on main script

* Updated container memory reading to: Mtotal-Mcached

* Fixed container memory read tests fixed

* Adapted container memory readings for ARM vs AMD64 docker versions

* Added common python NE version. Removed wiring pi download from installation as it is deprecated

* Dockerfile fix

* agent_main.py: use current Python 3

* docker-tools: use same base image as other components

+++ code/agent/agent_api.py
@@ -60,7 +60,7 @@ def local_peripheral_get_identifier(filepath):
     """
     try:
         peripheral_nuvla_id = NB.read_json_file(filepath)["id"]
-    except:
+    except (KeyError, FileNotFoundError):
         # if something happens, just return None
         return None
 
@@ -91,30 +91,36 @@ def post(payload):
     try:
         sanitize_peripheral_payload(payload)
     except KeyError as e:
-        logging.error("Payload {} is incomplete. Missing 'identifier'. {}".format(payload, e))
-        return {"error": "Payload {} is incomplete. Missing 'identifier'. {}".format(payload, e)}, 400
+        logging.error(f"Payload {payload} is incomplete. Missing 'identifier'. {e}")
+        return {"error": f"Payload {payload} is incomplete. Missing "
+                         f"'identifier'. {e}"}, 400
     except TypeError:
         logging.error("Payload {} malformed. It must be a JSON payload".format(payload))
-        return {"error": "Payload {} malformed. It must be a JSON payload".format(payload)}, 400
+        return {"error":f"Payload {payload} malformed. It must be a JSON payload"}, 400
 
     peripheral_identifier = payload['identifier']
     peripheral_filepath = "{}/{}".format(NB.peripherals_dir, peripheral_identifier)
 
     # Check if peripheral already exists locally before pushing to Nuvla
     if local_peripheral_exists(peripheral_filepath):
-        logging.error("Peripheral %s file already registered. Please delete it first" % peripheral_identifier)
-        return {"error": "Peripheral %s file already registered. Please delete it first" % peripheral_identifier}, 400
+        logging.error(f"Peripheral {peripheral_identifier} file already registered. "
+                      f"Please delete it first")
+        return {f"error": f"Peripheral {peripheral_identifier} file already registered."
+                          f" Please delete it first"}, 400
 
     # check if it already exists in Nuvla
     try:
-        existing_nuvla_per = NB.api().get('nuvlabox-peripheral', filter=f'identifier="{peripheral_identifier}"').data
+        existing_nuvla_per = NB.api().get(
+            'nuvlabox-peripheral',
+            filter=f'identifier="{peripheral_identifier}"').data
+
     except nuvla.api.api.NuvlaError as e:
         logging.exception("Unable to reach Nuvla")
         return e.response.json(), e.response.status_code
 
     if existing_nuvla_per.get('count', 0) > 0:
-        # already registered in Nuvla, but not locally...maybe something went wrong in a past mgmt action
-        # let's just update it
+        # already registered in Nuvla, but not locally...maybe something went wrong in a
+        # past mgmt action let's just update it
         fix_edit = modify(peripheral_identifier,
                           peripheral_nuvla_id=existing_nuvla_per['resources'][0]['id'],
                           payload=payload)
@@ -145,7 +151,7 @@ def post(payload):
         logging.info("Saving peripheral %s locally" % payload['id'])
         local_peripheral_save(peripheral_filepath, payload)
     except Exception as e:
-        logging.exception("Unable to save peripheral. Reverting request...")
+        logging.error("Unable to save peripheral. Reverting request...")
         modify(peripheral_identifier, peripheral_nuvla_id=payload['id'], action='DELETE')
         return {"error": "Unable to fulfill request: %s" % e}, 500
 
@@ -156,8 +162,8 @@ def modify(peripheral_identifier, peripheral_nuvla_id=None, action='PUT', payloa
     """ Modifies (edits or deletes) a peripheral from the local and Nuvla database
 
     :param peripheral_identifier: unique local identifier for the peripheral
-    :param peripheral_nuvla_id: (optional) Nuvla ID for the peripheral resource. If present, will not infer it from
-    the local file copy of the peripheral resource
+    :param peripheral_nuvla_id: (optional) Nuvla ID for the peripheral resource.
+    If present, will not infer it from the local file copy of the peripheral resource
     :param action: PUT or DELETE
     :param payload: body used for PUT
     :returns request message and status
@@ -170,41 +176,50 @@ def modify(peripheral_identifier, peripheral_nuvla_id=None, action='PUT', payloa
 
     peripheral_filepath = "{}/{}".format(NB.peripherals_dir, peripheral_identifier)
 
-    per_nuvla_id = peripheral_nuvla_id if peripheral_nuvla_id else local_peripheral_get_identifier(peripheral_filepath)
+    per_nuvla_id = peripheral_nuvla_id if peripheral_nuvla_id else \
+        local_peripheral_get_identifier(peripheral_filepath)
+
     if not per_nuvla_id:
-        logging.warning("{} not found and Nuvla resource ID not provided".format(peripheral_filepath))
+        logging.warning(f"{peripheral_filepath} not found and Nuvla resource ID not "
+                        f"provided")
         return {"error": "Peripheral not found"}, 404
 
     try:
         if action == 'DELETE':
-            out_peripheral = delete_peripheral(peripheral_nuvla_id, peripheral_filepath)
-            logging.info("Deleted {} from Nuvla".format(peripheral_nuvla_id))
+            out_peripheral = delete_peripheral(per_nuvla_id, peripheral_filepath)
+            logging.info(f"Deleted {per_nuvla_id} from Nuvla")
         else:
-            out_peripheral = edit_peripheral(peripheral_nuvla_id, payload, peripheral_filepath)
-            logging.info("Changed {} in Nuvla, with payload: {}".format(peripheral_nuvla_id, payload))
+            out_peripheral = edit_peripheral(per_nuvla_id, payload, peripheral_filepath)
+            logging.info(f"Changed {per_nuvla_id} in Nuvla, with payload: {payload}")
 
         return out_peripheral.data, out_peripheral.data.get('status', 200)
+
     except nuvla.api.api.NuvlaError as e:
-        if e.response.status_code == 404:
-            logging.warning(f"Peripheral {peripheral_nuvla_id} not found in Nuvla: {e.response.json()}")
+        if e.response.status_code == 404 and action == 'DELETE':
+            logging.warning(f"Peripheral {per_nuvla_id} not found in Nuvla: "
+                            f"{e.response.json()}")
 
-            if action == 'DELETE':
-                try:
-                    os.remove(peripheral_filepath)
-                except FileNotFoundError:
-                    pass
-                logging.info("Deleted {} from the NuvlaBox".format(peripheral_filepath))
-                return {"message": "Deleted %s" % peripheral_identifier}, 200
+            try:
+                os.remove(peripheral_filepath)
+            except FileNotFoundError:
+                pass
+            logging.info("Deleted {} from the NuvlaBox".format(peripheral_filepath))
+            return {"message": f"Deleted {peripheral_identifier}"}, 200
 
-        # Maybe something went wrong and we should try later, so keep the local peripheral copy alive
-        logging.warning("Cannot {} {} in Nuvla: {}".format(action, peripheral_nuvla_id, e.response.json()))
+        # Maybe something went wrong, and we should try later, so keep the local
+        # peripheral copy alive
+        logging.warning(f"Cannot {action} {peripheral_nuvla_id} in "
+                        f"Nuvla: {e.response.json()}")
         return e.response.json(), e.response.status_code
+
     except Exception as e:
-        logging.exception("While running {} on {} from Nuvla".format(action, peripheral_nuvla_id))
-        return {"error": f"Error occurred while doing {action} on {peripheral_identifier}: {str(e)}"}, 500
+        logging.exception(f"While running {action} on {peripheral_nuvla_id} from Nuvla")
+        return {"error": f"Error occurred while doing {action} on "
+                         f"{peripheral_identifier}: {str(e)}"}, 500
 
 
-def edit_peripheral(peripheral_id: str, body: dict, local_filepath: str) -> nuvla.api.api.CimiResponse:
+def edit_peripheral(peripheral_id: str, body: dict, local_filepath: str) \
+        -> nuvla.api.api.CimiResponse:
     """
     Edits a peripheral in Nuvla and locally if needed
 
@@ -213,14 +228,15 @@ def edit_peripheral(peripheral_id: str, body: dict, local_filepath: str) -> nuvl
     :param local_filepath: local peripheral file path
     :return: Nuvla response to the edit request
     """
-    r = NB.api().edit(peripheral_id, body)
+    response = NB.api().edit(peripheral_id, body)
     if local_peripheral_exists(local_filepath):
         local_peripheral_update(local_filepath, body)
 
-    return r
+    return response
 
 
-def delete_peripheral(peripheral_id: str, local_filepath: str) -> nuvla.api.api.CimiResponse:
+def delete_peripheral(peripheral_id: str, local_filepath: str) \
+        -> nuvla.api.api.CimiResponse:
     """
     Deletes a peripheral from Nuvla and locally as well
 
@@ -230,11 +246,11 @@ def delete_peripheral(peripheral_id: str, local_filepath: str) -> nuvla.api.api.
 
     Returns: Nuvla response to the delete request
     """
-    r = NB.api().delete(peripheral_id)
+    response = NB.api().delete(peripheral_id)
     if local_peripheral_exists(local_filepath):
         os.remove(local_filepath)
 
-    return r
+    return response
 
 
 def find(parameter, value, identifier_pattern):
@@ -242,13 +258,14 @@ def find(parameter, value, identifier_pattern):
 
     :param parameter: name of the parameter to search for
     :param value: value of that parameter
-    :param identifier_pattern: regex expression to limit the search query to peripherals matching the identifier pattern
+    :param identifier_pattern: regex expression to limit the search query to peripherals
+    matching the identifier pattern
     :returns list of peripheral matching the search query
     """
     matched_peripherals = {}
 
-    search_dir = "{}/{}".format(NB.peripherals_dir, identifier_pattern) if identifier_pattern \
-        else NB.peripherals_dir + "/**/**"
+    search_dir = "{}/{}".format(NB.peripherals_dir, identifier_pattern) \
+        if identifier_pattern else NB.peripherals_dir + "/**/**"
 
     for filename in glob.iglob(search_dir, recursive=True):
         if os.path.isdir(filename):
@@ -257,12 +274,13 @@ def find(parameter, value, identifier_pattern):
         with open(filename) as f:
             try:
                 content = json.loads(f.read())
-            except:
+            except (json.JSONDecodeError, IOError):
                 continue
 
         if parameter and value:
             if parameter in content and content[parameter] == value:
-                matched_peripherals[filename.replace(f'{NB.peripherals_dir}/', '')] = content
+                matched_peripherals[filename.replace(f'{NB.peripherals_dir}/', '')] = \
+                    content
         else:
             matched_peripherals[filename.replace(f'{NB.peripherals_dir}/', '')] = content
 
@@ -270,9 +288,10 @@ def find(parameter, value, identifier_pattern):
 
 
 def get(identifier):
-    """ Finds a specific locally registered peripherals that matches the identifier, by filename
+    """ Finds a specific locally registered peripherals that matches the identifier, by
+    filename
 
-    :param identifier: peripheral identifier and filename (including its subfolder if any)
+    :param identifier: peripheral identifier and filename(including its sub folder if any)
     :returns peripheral content
     """
     search_for = "{}/{}".format(NB.peripherals_dir, identifier)
@@ -281,7 +300,7 @@ def get(identifier):
         with open(search_for) as p:
             try:
                 return json.loads(p.read()), 200
-            except:
+            except (json.JSONDecodeError, IOError):
                 return {"error": "Cannot read peripheral information"}, 500
     else:
         return {"error": "Peripheral not found"}, 404
@@ -294,8 +313,8 @@ def save_vpn_ip(ip):
     :param ip: string
     :return:
     """
-    with open(NB.vpn_ip_file, 'w') as vpnip:
-        vpnip.write(str(ip))
+    with open(NB.vpn_ip_file, 'w') as vpn_ip:
+        vpn_ip.write(str(ip))
 
 
 def save_vulnerabilities(vulnerabilities):
@@ -305,5 +324,5 @@ def save_vulnerabilities(vulnerabilities):
     :param vulnerabilities: as JSON
     :return:
     """
-    with open(NB.vulnerabilities_file, 'w') as vf:
-        vf.write(json.dumps(vulnerabilities))
+    with open(NB.vulnerabilities_file, 'w') as file:
+        file.write(json.dumps(vulnerabilities))
